restructuring, add api for Acn Peers info
